Moved ML-KEM to the FIPS module
