+++ crypto/fipsmodule/ml_kem/ml_kem_ref/verify.c
@@ -2,7 +2,7 @@
 #include <stdint.h>
 #include "verify.h"
 
-#include "../../internal.h"
+#include "../../../internal.h"
 
 /*************************************************
 * Name:        verify
